Browser Tests

+++ sdk/tables/data-tables/test/public/accessPolicy.spec.ts
@@ -35,7 +35,7 @@ describe(`Access Policy operations`, () => {
   });
 
   after(async () => {
-    if (!isPlaybackMode()) {
+    if (!isPlaybackMode() && isNode) {
       await unrecordedClient.deleteTable();
     }
   });
